+++ CHANGELOG.md
@@ -7,6 +7,11 @@
         - Use div for inspector form wrapped extra questions. #3250
     - Admin improvements:
         - Enable per-category hint customisation.
+        - Move ban/unban buttons to user edit admin page.
+        - Add link to user edit admin from report/update edit admin.
+        - Improve layout of some admin pages.
+    - Security:
+        - Increase minimum password length to eight.
 
 * v3.1 (16th November 2020)
     - Security: